requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pytz==2023.3.post1
popbill==1.60.1
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6


